+++ src/cli/commands/public-cmds/import-cmd.js
@@ -10,13 +10,15 @@ import { ComponentWithDependencies } from '../../../scope';
 import type { ImportOptions, ImportDetails } from '../../../consumer/component-ops/import-components';
 import type { EnvironmentOptions } from '../../../api/consumer/lib/import';
 import GeneralError from '../../../error/general-error';
-import { BASE_DOCS_DOMAIN } from '../../../constants';
+import { BASE_DOCS_DOMAIN, WILDCARD_HELP } from '../../../constants';
 import { MergeOptions } from '../../../consumer/versions-ops/merge-version/merge-version';
 import type { MergeStrategy } from '../../../consumer/versions-ops/merge-version/merge-version';
 
 export default class Import extends Command {
   name = 'import [ids...]';
-  description = `import components into your current workspace.\n  https://${BASE_DOCS_DOMAIN}/docs/importing-components.html`;
+  description = `import components into your current workspace.
+  https://${BASE_DOCS_DOMAIN}/docs/importing-components.html
+  ${WILDCARD_HELP('import')}`;
   alias = 'i';
   opts = [
     ['t', 'tester', 'import a tester environment component'],
@@ -32,6 +34,7 @@ export default class Import extends Command {
     ['d', 'display-dependencies', 'display the imported dependencies'],
     ['O', 'override', 'override local changes'],
     ['v', 'verbose', 'showing verbose output for inspection'],
+    ['j', 'json', 'return the output as JSON'],
     ['', 'ignore-dist', "skip writing the component's build files during import"],
     [
       '',
@@ -71,6 +74,7 @@ export default class Import extends Command {
       environment = false,
       override = false,
       verbose = false,
+      json = false,
       ignoreDist = false,
       conf,
       skipNpmInstall = false,
@@ -88,6 +92,7 @@ export default class Import extends Command {
       environment?: boolean,
       override?: boolean,
       verbose?: boolean,
+      json?: boolean,
       ignoreDist?: boolean,
       conf?: string,
       skipNpmInstall?: boolean,
@@ -141,9 +146,11 @@ export default class Import extends Command {
     if (typeof conf === 'string') {
       importOptions.configDir = conf;
     }
-    return importAction(environmentOptions, importOptions, packageManagerArgs).then(importResults =>
-      R.assoc('displayDependencies', displayDependencies, importResults)
-    );
+    return importAction(environmentOptions, importOptions, packageManagerArgs).then(importResults => ({
+      displayDependencies,
+      json,
+      ...importResults
+    }));
   }
 
   report({
@@ -151,7 +158,8 @@ export default class Import extends Command {
     envComponents,
     importDetails,
     warnings,
-    displayDependencies
+    displayDependencies,
+    json
   }: {
     dependencies?: ComponentWithDependencies[],
     envComponents?: Component[],
@@ -161,8 +169,12 @@ export default class Import extends Command {
       notInNodeModules: [],
       notInBoth: []
     },
-    displayDependencies?: boolean
+    displayDependencies: boolean,
+    json: boolean
   }): string {
+    if (json) {
+      return JSON.stringify({ importDetails, warnings }, null, 4);
+    }
     let dependenciesOutput;
     let envComponentsOutput;